+++ xl_auth/oauth/grant/__init__.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-"""The grant module."""
+"""The OAuth Grant module."""
 
 from __future__ import absolute_import, division, print_function, unicode_literals
